Issue Fix: fix function test 'RenameFileBeforeClose' failure
RootCause: the previous logic of FileMetaDataManager::Rename would
do nothing is the file with new name exist. This is not right as it
will cause the file metadata used the existing one which is uncorrect
rename behaviour. For the function test case, an existing new file
has empty content and the file to be renamed has non empty content,
so the rename will truncate the file to empty, this cause the follwing
content validation fail.
Resolution: if an file with target new filename exist, remove it.
This will enable the following replacement happen.