Fake the chmod and chown with null implementation.
We fake the meta related api (such as chmod, chown, utimens) to enable
some app works on a certain case, for example, rsync and nfs client
write operation like cp.
1. rsync always call mktemp to create a temp file then call chmod(0644)
2. nfs client cp call fuse mknod then create a file with 10000 mode
then call chmod(10644)
To enable these case work we can ignore chmod request by fake it
with a null implementation. And qsfs can specify the fs file/dir mode
and the uig/gid, so this should be work if we do not care the
permission bit.