Add option filesystem capacity.
Filesystem capcity is used to spcecify the filesystem size, it will
be displayed when use command df to report filesystem disk space usage.
This value not means the true capacity as object storage has a limiteless
scabality.
The default value is 1PB, it can be specified to be numbers of GB. And
if the used size surpass this specified value, the capacity will be ajust
to the same value as used size, in this case, the use will 100% but you
can still store more data along with a increased capacity.